services/pipeline/requirements.txt
numpy
torch
torchvision
pillow
opencv-python
rembg
onnxruntime
transformers
ultralytics
